pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xaudit"
version = "0.1.0"
description = "Reasoning quality analysis for autonomous systems"
authors = [{name = "Aditya Khankar"}]
license = {text = "MIT"}
requires-python = ">=3.11"
readme = "README.md"
keywords = [
    "ai",
    "agents",
    "evaluation",
    "behavioral-economics",
    "llm",
    "langchain",
    "autonomous-agents",
    "agent-evals",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "matplotlib>=3.8.0",
    "pyfiglet>=0.8.post1",
]

[project.urls]
Homepage = "https://github.com/Aditya-Khankar/xaudit"
Repository = "https://github.com/Aditya-Khankar/xaudit"
Issues = "https://github.com/Aditya-Khankar/xaudit/issues"

[project.scripts]
xaudit = "xaudit.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
generate = [
    "langchain-core>=0.1.0",
    "langgraph>=0.1.0",
    "google-genai>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["xaudit*"]

[tool.setuptools.package-data]
xaudit = ["sample_traces/*.json"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
